+++ src/main/java/com/example/demo/controller/KakaoSearchController.java
@@ -1,6 +1,6 @@
 package com.example.demo.controller;
 
-import com.example.demo.domain.SearchBook;
+import com.example.demo.object.SearchBook;
 import com.example.demo.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
