--- src/main/java/com/example/demo/domain/BestSeller.java
@@ -1,135 +1,0 @@
-package com.example.demo.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-public class BestSeller {
-    private Long itemId;
-    private String title;
-    private String description;
-    private String pubDate;
-    private String coverSmallUrl;
-    private String coverLargeUrl;
-    private int categoryId;
-    private String categoryName;
-    private String publisher;
-    private Long customerReviewRank;
-    private String author;
-    private String translator;
-    private Long isbn;
-    private int rank;
-
-    public Long getItemId() {
-        return itemId;
-    }
-
-    public void setItemId(Long itemId) {
-        this.itemId = itemId;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getPubDate() {
-        return pubDate;
-    }
-
-    public void setPubDate(String pubDate) {
-        this.pubDate = pubDate;
-    }
-
-    public String getCoverSmallUrl() {
-        return coverSmallUrl;
-    }
-
-    public void setCoverSmallUrl(String coverSmallUrl) {
-        this.coverSmallUrl = coverSmallUrl;
-    }
-
-    public String getCoverLargeUrl() {
-        return coverLargeUrl;
-    }
-
-    public void setCoverLargeUrl(String coverLargeUrl) {
-        this.coverLargeUrl = coverLargeUrl;
-    }
-
-    public int getCategoryId() {
-        return categoryId;
-    }
-
-    public void setCategoryId(int categoryId) {
-        this.categoryId = categoryId;
-    }
-
-    public String getCategoryName() {
-        return categoryName;
-    }
-
-    public void setCategoryName(String categoryName) {
-        this.categoryName = categoryName;
-    }
-
-    public String getPublisher() {
-        return publisher;
-    }
-
-    public void setPublisher(String publisher) {
-        this.publisher = publisher;
-    }
-
-    public Long getCustomerReviewRank() {
-        return customerReviewRank;
-    }
-
-    public void setCustomerReviewRank(Long customerReviewRank) {
-        this.customerReviewRank = customerReviewRank;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getTranslator() {
-        return translator;
-    }
-
-    public void setTranslator(String translator) {
-        this.translator = translator;
-    }
-
-    public Long getIsbn() {
-        return isbn;
-    }
-
-    public void setIsbn(Long isbn) {
-        this.isbn = isbn;
-    }
-
-    public int getRank() {
-        return rank;
-    }
-
-    public void setRank(int rank) {
-        this.rank = rank;
-    }
-}
