+++ src/main/java/com/example/demo/object/SearchBook.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.object;
 
 import lombok.Data;
