+++ src/main/resources/application.properties
@@ -1,19 +1,19 @@
-logging.level.org.apache.coyote.http11=debug
-logging.level.org.hibernate.type.descriptor.sql=trace
+#logging.level.org.apache.coyote.http11=debug
+#logging.level.org.hibernate.type.descriptor.sql=trace
 
 # MySQL Driver
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # DB URL
-spring.datasource.url=jdbc:mysql://localhost:3306/mydb?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul
+spring.datasource.url=jdbc:mysql://localhost:3306/book?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul
 
 
 # DB username
-spring.datasource.username=juae
+spring.datasource.username=root
 
 # DB password
 
-spring.datasource.password=wndorn01
+spring.datasource.password=1234
 
 # ??? SQL ?? ??
 spring.jpa.show-sql=true
