+++ src/main/resources/static/asset-manifest.json
@@ -1,23 +1,23 @@
 {
   "files": {
-    "main.css": "/static/css/main.79363002.chunk.css",
-    "main.js": "/static/js/main.5523653a.chunk.js",
-    "main.js.map": "/static/js/main.5523653a.chunk.js.map",
-    "runtime-main.js": "/static/js/runtime-main.a2b138b1.js",
-    "runtime-main.js.map": "/static/js/runtime-main.a2b138b1.js.map",
-    "static/js/2.75202d2e.chunk.js": "/static/js/2.75202d2e.chunk.js",
-    "static/js/2.75202d2e.chunk.js.map": "/static/js/2.75202d2e.chunk.js.map",
-    "static/js/3.2b0f760f.chunk.js": "/static/js/3.2b0f760f.chunk.js",
-    "static/js/3.2b0f760f.chunk.js.map": "/static/js/3.2b0f760f.chunk.js.map",
+    "main.css": "/static/css/main.c251457a.chunk.css",
+    "main.js": "/static/js/main.0c87598c.chunk.js",
+    "main.js.map": "/static/js/main.0c87598c.chunk.js.map",
+    "runtime-main.js": "/static/js/runtime-main.ed6fa33d.js",
+    "runtime-main.js.map": "/static/js/runtime-main.ed6fa33d.js.map",
+    "static/js/2.871d9e05.chunk.js": "/static/js/2.871d9e05.chunk.js",
+    "static/js/2.871d9e05.chunk.js.map": "/static/js/2.871d9e05.chunk.js.map",
+    "static/js/3.95f28b19.chunk.js": "/static/js/3.95f28b19.chunk.js",
+    "static/js/3.95f28b19.chunk.js.map": "/static/js/3.95f28b19.chunk.js.map",
     "index.html": "/index.html",
-    "static/css/main.79363002.chunk.css.map": "/static/css/main.79363002.chunk.css.map",
-    "static/js/2.75202d2e.chunk.js.LICENSE.txt": "/static/js/2.75202d2e.chunk.js.LICENSE.txt",
+    "static/css/main.c251457a.chunk.css.map": "/static/css/main.c251457a.chunk.css.map",
+    "static/js/2.871d9e05.chunk.js.LICENSE.txt": "/static/js/2.871d9e05.chunk.js.LICENSE.txt",
     "static/media/App.css": "/static/media/Pretendard-Medium.59b075f0.woff2"
   },
   "entrypoints": [
-    "static/js/runtime-main.a2b138b1.js",
-    "static/js/2.75202d2e.chunk.js",
-    "static/css/main.79363002.chunk.css",
-    "static/js/main.5523653a.chunk.js"
+    "static/js/runtime-main.ed6fa33d.js",
+    "static/js/2.871d9e05.chunk.js",
+    "static/css/main.c251457a.chunk.css",
+    "static/js/main.0c87598c.chunk.js"
   ]
 }
